Fix flaky datetime tests: increase paid_at delta and cast diffInDays to int

+++ tests/unit/SubscriptionTest.php
@@ -187,7 +187,7 @@ class SubscriptionTest extends TestCase
         $subscription = Subscription::factory()->active()->create([
             'expires_at' => Carbon::parse('+ 3 weeks')
         ]);
-        $this->assertEqualsWithDelta(Carbon::now()->diffInDays(Carbon::parse('+ 3 weeks')), $subscription->remaining_days, 1);
+        $this->assertEqualsWithDelta((int) Carbon::now()->diffInDays(Carbon::parse('+ 3 weeks')), $subscription->remaining_days, 1);
 
         $expiredSubscription = Subscription::factory()->expired()->create();
         $this->assertEquals(0, $expiredSubscription->remaining_days);
